pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "Email triage with Claude AI"
authors = [
    {name = "Lenna"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=1.2.20",
    "python-dotenv>=1.2.2"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
